Move schema presenter into own namespace
